Rename namespaces. Remove main repo of data
