+++ DataContext.Sqlite/DataContext.Sqlite.csproj
@@ -11,7 +11,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\EntityContext.Models\EntityContext.Models.csproj" />
+    <ProjectReference Include="..\DataContext.Models\DataContext.Models.csproj" />
+    <ProjectReference Include="..\EntityContext.Models\DataContext.Models.csproj" />
   </ItemGroup>
 
 </Project>
