--- EntityContext/Repositories/IBoxCommonRepository.cs
@@ -1,13 +1,0 @@
-﻿using EntityContext.Models.Models;
-
-namespace EntityContext.Repositories;
-
-public interface IBoxCommonRepository
-{
-    Task<BoxEfModel?> CreateAsync(BoxEfModel box);
-    Task<IEnumerable<BoxEfModel>> RetrieveAllAsync();
-    //Task<IList<BoxEfModel>> RetrieveAllAsync();
-    Task<BoxEfModel?> RetrieveAsync(string id);
-    Task<BoxEfModel?> UpdateAsync(string id, BoxEfModel box);
-    Task<bool?> DeleteAsync(string id);
-}
