--- PalletStorage.Common.Controllers/PalletStorage.Common.Controllers.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\DataContext.Sqlite\DataContext.Sqlite.csproj" />
-    <ProjectReference Include="..\EntityContext\EntityContext.csproj" />
-    <ProjectReference Include="..\PalletStorage.Common\PalletStorage.Common.csproj" />
-    <ProjectReference Include="..\PalletStorage.WebApi.Models\PalletStorage.WebApi.Models.csproj" />
-  </ItemGroup>
-
-</Project>
