+++ PalletStorage.Repositories/PalletStorage.Repositories.csproj
@@ -7,8 +7,9 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\DataContext.Models\DataContext.Models.csproj" />
     <ProjectReference Include="..\DataContext.Sqlite\DataContext.Sqlite.csproj" />
-    <ProjectReference Include="..\EntityContext.Models\EntityContext.Models.csproj" />
+    <ProjectReference Include="..\EntityContext.Models\DataContext.Models.csproj" />
   </ItemGroup>
 
 </Project>
