+++ PalletStorage.Tests/PalletStorage.Tests.csproj
@@ -24,9 +24,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\DataContext.Models\DataContext.Models.csproj" />
     <ProjectReference Include="..\DataContext.Sqlite\DataContext.Sqlite.csproj" />
-    <ProjectReference Include="..\PalletStorage.Common.Controllers\PalletStorage.Common.Controllers.csproj" />
     <ProjectReference Include="..\PalletStorage.Common\PalletStorage.Common.csproj" />
+    <ProjectReference Include="..\PalletStorage.Repositories\PalletStorage.Repositories.csproj" />
   </ItemGroup>
 
 </Project>
