+++ PalletStorage.WebApi/Program.cs
@@ -1,9 +1,9 @@
 using DataContext.Sqlite;
-using EntityContext.Repositories;
 using Microsoft.OpenApi.Models;
 using Microsoft.AspNetCore.Mvc.Formatters;
 using Swashbuckle.AspNetCore.SwaggerUI; // SubmitMethod
 using Microsoft.AspNetCore.HttpLogging;
+using PalletStorage.Repositories.Repositories;
 using PalletStorage.WebApi.Repositories;
 // HttpLoggingFields
